+++ src/View/AttributeFormatterTrait.php
@@ -15,21 +15,22 @@ trait AttributeFormatterTrait {
 
 		$html = [];
 		foreach ( $attributes as $key => $value ) {
+
 			if ( is_bool( $value ) ) {
 				if ( $value ) {
 					$html[] = $this->esc_attr( $key );
 				}
-				continue;
-			}
+			} else {
 
-			if ( is_array( $value ) ) {
-				$value = json_encode( $value );
+				if ( is_array( $value ) ) {
+					$value = json_encode( $value );
+				}
+				$html[] = sprintf(
+					'%s="%s"',
+					$this->esc_attr( $key ),
+					$this->esc_attr( $value )
+				);
 			}
-			$html[] = sprintf(
-				'%s="%s"',
-				$this->esc_attr( $key ),
-				$this->esc_attr( $value )
-			);
 		}
 
 		return implode( ' ', $html );