+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
 
+## 0.2.0 (NOT RELEASED)
+- Added new tests for `AttributeFormatterTrait`.
+- Implemented boolean-attributes like `disabled` or `required`.
+- Introduced new method `is_disabled` to `ElementInterface`.
+- Don't bind data to Elements or validate Elements which are disabled. 
+
 ## 0.1.0
 - First release